Remove the !reminder-timestamp file if it's no longer needed
If we didn't have to ignore any files because there is no !ready, we no
longer need the timestamp of the last reminder about that, so we can remove
the !reminder-timestamp file.

This means that for a given maintainer, if another instance of missing
!ready occurs, the reminder is issued immediately, rather than being delayed
until the time the reminder for the first instance was scheduled.