leads_email should be empty unless something went wrong
Add AbeyanceHandler(), which stores log messages and discards them if a
severity threshold is not crossed.

Use that with a threshold of ERROR for leads_mail (so it is silent unless
errors occured) and INFO for maintainer_mail (so it reports file moves).

Also, move mail_logs to a more logical place