Removing setup.hint files isn't allowed
Removing a package doesn't make much sense currently.

We can't remove a package when it has been replaced by another one, all we
can do is replace it with an empty, obsolete one which depends on it's
replacement, so existing installs are updated to that.

Removing a package which is no longer relevant (i.e. an old soversion of a
shared library which has no users left) is permissible, but the files it
contains will linger forever in any existing installs.