Better checking for corrupt packages
Log an error if we need to access a package file in the release area which
isn't a valid compressed archive.  Ideally, we'd use tarfile.is_tarfile() to
check that, but that can propagate exceptions from the decompressor.

(Perhaps we should also ignore these invalid files, but I'm not sure that is
a good idea.)

Note that we avoid opening the package files in the release area unless
unavoidable, as it's expensive.  Validating a tar file involves reading the
entire file.

Always validate uploaded package files, and reject attempts to upload such
invalid files

Add appropriate tests