Move stale package versions to the vault
Move stale package versions to the vault.  A package version is not stale if
any of the following is true:

* The version is used as curr:, prev: or test: (due to either being
explictly listed in override.hint or implictly chosen)
* The version is listed as keep: in override.hint
* It's one of the highest n versions
* The first version which is less than t days old, and all higher versions.

where:
 n is given by keep-count: in override.hint
 t is given by keep-days: in override.hint

These tunables currently default to 0, so no more packages are kept than
would be kept by stalepkgs, although these defaults might change in future.

We track the freshness or staleness of install and source packages
separately.  Unlike stalepkgs, a fresh empty install file does not cause the
corresponding source package to be considered fresh.  Also remove any
pvr.hint which is stale.

Add the --no-stale option to disable.

First we check for stale pkgs in release area.  stale pkgs arising due to
more fresh pckages being uploaded are reported as a part of the upload.

Avoid doing unnecessary work if nothing is stale.

Update tests