Report conflicting movelists as an error
We probably want to told when we are uploading a package which is
immediately going to be automatically removed as stale.

Uploading a package at the same time as a marking it for removal doesn't
make any sense, so report that as an error.

Update per-version testpackage to avoid this error.