Add support for explicit 'test:' in pvr.hint
With hindsight, override.hint seems to have been a bit of a mistake, as it
suffers from the same problem as setup.hint, that a careless update to it
unintentionally affects other versions.

So, as an alternate mechanism, add support for an explicit 'test:' in
pvr.hint, marking a version as a test version.

If multiple versions have are marked thus, the highest version will be
assigned to the test stability level. Versions marked as test cannot be
assigned to any other stability level.