Add a disable-check: key in pvr.hint to disable certain checks
Add a disable-check: key in pvr.hint to disable certain checks, to gradually
supersede the static list in past_mistakes (which needs calm restarting to
re-read)

disable-check: contains a list of tokens, which indicates which checks are
to be downgraded from errors to warnings

in pvr.hint, disable-check: can contain:

  empty-obsolete
     check that if a binary package is empty and has no source package, it
     is marked obsolete

  unique-version
     check that this binary package has the same version as all other
     packages made from the same source package

These disable-check:s end up being applied to new versions when something is
split out from an existing source.  That kind of suggests something is not
quite right about the test we are doing here.  But we are limited because we
don't have a source-centric view of things due to source packages not really
being a thing yet...