Allow the effective version to be overridden by version: hint
The archive filenames appear in setup.ini unchanged, but the version is
taken from the hint, rather than deduced from the archive filename.

This requires a change to collecting tars per VR, rather than keeping them
all in one list.

This could be used as a way to fix the epoch of packages which have a
decreased version number, but were built without bumping the epoch.

This could also be used to have a properly ordered version number whilst
using upstream version numbers which don't do that for the archive
filenames.