Add a replace-versions: key in override.hint to list versions which should always be replaced
This is intended for (temporarily) dealing with anomalous versions e.g.
those which have been removed (or relabelled as test), without a
superseding higher version existing, and indicates to setup that the named
versions(s), if installed, should be replaced with different version.

Also warn if this key pointlessly lists versions which will be replaced by
the current version anyhow.