Only update package metadata dump if it's contents will have changed
Even if we were to drop the timestamp element from the dump, we can't
write the json and then check if it's changed, as the order of python
lists and dict keys is non-deterministic.

So instead, just write it when it seems that something has changed.

This allows fetching the dump with If-Modified-Since: to work as
expected.