Don't use defaultdict for vermap
Written as it is, this needs more care to avoid accidentally creating
version of packages.

Note that after this, we have two different definitions of the set of
versions for a package: vermap.keys() and version_hints.keys().  One is
the set of versions for which we currently have archives (after
vaulting), the other, the set of versions for which we've ever seen
hints.