Make keep-count: not retain test: versions
(Note that we will always retain the version assigned to the test:
stability level (the highest test: version, unless specified in
override.hint), if any.

For the moment, also retain 2 test versions, so the change in
behaviour isn't large. This could be tuned down to retaining 1 test
version in future.

v2:
Handle keep-count:0 correctly