Make --force work more sensibly when daemonized
If daemonized, --force should force regeneration of static content in
htdocs initially (in case the generation code has changed), but that
static content should only by updated as needed on subsequent loops.

Use open_amifc() so that mtimes aren't changed by forced regeneration of
static content, if it hasn't actually changed.

(This also atomically updates that content, just in case someone happens
to read it while it's being updated)

Regenerating package listing pages is expensive (since we have to read
all of every tar archive to do so), so only do that with '--force --force'.