Remove support for curr: appearing in override.hint
This doesn't really work the way it's expected to across setup versions,
and isn't a good way of approaching the problem (current alternatives:
make the version you want to be preferred have the highest non-test
version, or use replace-versions: to cause higher versions to be
uninstalled).

Future work: lots of cruft for handling stability level overrides
remains, which can all be cleaned up.