Bump the minimum setup version
We start to raise the minimum setup version, to hopefully cause people
to stop using those old versions, in anticipation of dropping signing
with the key those old version expect.

Unfortunately, this doesn't help with the very old versions which don't
support setup-minimum-version:. which ~1% of IP activity is still using.