Improve empty package checks
An empty install package without depends should be in the '_obsolete'
category.

(We're now checking all install packages, not just ones without
external-source: (which we ignored here for somewhat bogus reasons), so
add additional exceptions to past_mistakes.)

An empty install package with depends should be in the 'virtual'
category (but there are many old obsoletion packages which look like
that, so allow '_obsolete' as well).

Also check for empty source packages (and then fix the empty rpm-doc-src
we have in test data).