Drop validating versions appearing in override.hint keys
This is a followup to d863c1a2, 3cc08991 and 27852b24.

Since the curr:, prev:, test: keys are no longer permitted in
override.hint, there's no point in checking their value.

Do some simplifcation of the code which assigns versions to stability
levels, but there's probably more to do there.