Improve handling of old-style obsoletion packages
Identify old-style obsoletion packages (where the package is empty, has
a category of _obsolete, and requires: it's replacement), and upgrade
them to the new style (where the obsoleting package has an obsoletes:).

This is done per-arch, simply because lots of these historic obsoletion
packages only exist for x86, and it's redundant to obsolete: them on
x86_64.

Treating this similarly to past_mistakes.missing_obsoletes means that a
warning occurs when new version of package is missing that expected
obsolete:

Initially only apply to packages over a certain age, so we can observe
the effect on a small number of packages to make checking it's doing the
right thing easier.