Ignore 'skip' hint
Ignore the 'skip' hint, as it's meaningless, now we properly classify
packages as source or install.

Instead, if installing a non-obsolete package will do nothing, mark it
as not_for_output.

Future work: also allow the 'does nothing' check to apply to '_obsolete'
packages. At the moment this would catch a few more packages which
probably need to be obsoleted: by something before they can be safely
removed.

(It's not safe to simply vault them as we need to ensure any existing
installed files owned by that package are removed when a system with
them installed is updated.)