Expire superseded test versions
Expire superseded test versions (that is: expire test versions where a
non-test version with a higher version number exists), unless disabled by
the 'keep-superseded-test' hint.

(For this to work as expected, don't unconditionally mark the version
assigned to the test: stability level as not to be expired, so it can be
expired if superseded).