Drop computing and using 'prev' stability level
It's not used for anything.

Follow up to d8869d2e and 3cc08991.

The whole stability level concept can be ripped out, since there's no
way to set them anymore, so they are just implicit from version sorting.

This stops retaining an additional version when 'keep-count: 1' is
requested, but that is what's being asked for (only GeoIP-database and
clamav-db-daily currently).