Expire unused, deprecated, old shared library versions
Expire shared library versions which are:
- unused: no packages depend on it (rdepends is empty)
- deprecated: a later soversion exists (or the solibrary is no longer generated by the source)
- old: more than some number [1] of years old

girepository packages are 1:1 correspondence with the solib they are a
binding for, so shouldn't be cause of retention either (but don't
include them in the deprecated soversions report)

[1] initally 10 years, which is sufficently far back not to catch any
packages.  We will wind that forward gradually, so we can observe the
effect on a small number of packages initially.