Separate collect and read package files steps
This is necessary when different versions of a src package might exist
in both arch/ and src/ paths.

Update test data for change to Package() repr, as we no longer have
pkgpath, just then package name

This is preparatory work for allowing a couple of things: putting new
source packages in src/ (without moving the existing ones under arch/),
re-structuring the data model so that arch is inside package, rather
than containing a collection of them.

This improves package authorization so it's possible for a package can
replace parts of another package (without needing rearrangement of the
package locations), by requiring authorization for *all* paths at which
a package exists.

Change to using a class rather than named tuple to hold scan results, so
it can be mutable in uploads.auth_check()

Also: drop no longer needed use of a defaultdict for holding packages in
uploads.scan()