Revise maintainers module
The data ordered by package name is the one we use the most often, so
make that the primary form.

Also use mtime_cache for the result of parsing cygwin-pkg-maint.

v2:
Fix mkgitolite and reports

Keep maintainers in a list, not a set, because we need "first named
maintainer".

Don't write permissions lines with an empty list of names, as that's
syntactically incorrect.