Emit source packages and 'srcpkg:' references to them in setup.ini
Emit source packages (their names end with '-src') into setup.ini

Emit 'srcpkg:' lines referencing them.

setup 2.939 and later will interpret 'srcpkg:' lines, which are ignored
as unrecongized by earlier versions of setup.

We also continue to emit the existing 'source:' lines for the benefit of
older versions of setup.

setup 2.939 and later can handle the presence of both 'source:' and
'srcpkg:'. The ordering is not signficant since the presence of
'srcpkg:' takes priority.

The ordering of a package and it's corresponding source package is not
significant since they are independent packages to setup.