Rework how 'OBSOLETE' is handled in package list
Previously, we'd do some rearrangement of the release area, so the
obsoleted package would be moved under the obsoleting package, so it
didn't need to exist at all in the package list and we could safely
ignore it.

We don't need or want to bother with that now, so let obsolete packages
paths exist.

Future work: This needs some thought about how we want to work.  Perhaps
prohibit uploads?  Perhaps error if the package marked OBSOLETE isn't
actually obsoleted by something?