Add the ability to ingest arch-tagged package filenames
Change to deriving the name of the listing file from package name and
version, rather than from the tar archive filename, because that would
include the arch tag.

That exposes an interesting anomaly: We are ignoring any version:
override when generating the listing filename. Since it's taken into
account where the link is made to that file on the package summary page,
that seems like a bug which this change accidentally fixes.

Adjust the test data to reflect this fix.