Improve reliability of detecting upload available.
Detecting a !ready file being placed into the upload directory is
problematic, due to possible races in setting watches on a just created
directory.

So instead, we just watch for a file touched when the sftp session ends.

This means we more reliably detect when a !ready file has been uploaded,
at the cost of doing extra scans (in the probably rare case) when a sftp
session hasn't uploaded a !ready file.