Improve maintainer last activity estimate
Ignore timestamp of co-maintained packages when estimating last date of
maintainer activity.  This helps the inactive mainainter tool not to be
so easily fooled by co-maintainership.

(This is still not perfect, but... well, eventually we'll be in a world
where we have ssh activity timestamps for all maintainers and don't have
to resort to guessing based on package timestamps)