Properly ignore changes in staging tmpdir
fetcher makes a temporary dir (with mkdtemp, so it's probably called
'tmpXXXXXXXX') under 'tmp' in the stagingdir, so to properly ignore
changes there we should do so if the path ends with tmp, not the
filename is temp.

This has caused some problems with racyness, as we start trying to
process the staging dir before the upload has been moved into place.